+++ lua/plugins/lsp.lua
@@ -51,9 +51,9 @@ local function configure_lua(lspconfig, capabilities)
       -- access workspace path
       local path = client.workspace_folders[1].name
       if
-        -- there is no workspace level config for lua-language-server
-        not vim.loop.fs_stat(path .. "/.luarc.json")
-        and not vim.loop.fs_stat(path .. "/.luarc.jsonc")
+      -- there is no workspace level config for lua-language-server
+          not vim.loop.fs_stat(path .. "/.luarc.json")
+          and not vim.loop.fs_stat(path .. "/.luarc.jsonc")
       then
         -- setup server for neovim and config editing
         client.config.settings = vim.tbl_deep_extend("force", client.config.settings, {
@@ -97,6 +97,11 @@ local function configure_python(lspconfig, capabilities)
     on_attach = function(_, bufnr)
       on_attach(bufnr)
     end,
+    settings = {
+      python = {
+        pythonPath = settings.pyright.pythonPath or vim.fn.exepath("python")
+      }
+    },
   })
 end